changed average return type

+++ src/main/java/chartcalc/server/dispersion/dto/DispersionResponseDto.java
@@ -16,5 +16,5 @@ public class DispersionResponseDto {
 
 	Integer interval;
 
-	Integer averageReturn;
+	Double averageReturn;
 }
